+++ src/PAK-Samples/MapReferenceToPaper/packages.config
@@ -1,5 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="FuzzyString" version="1.0.0" targetFramework="net462" />
   <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net462" />
 </packages>
